Added the first version of the model. This has been "free hand tuning" only, i.e. parameters were tuned so as to qualitatively produce behaviour I was aware of from my understanding of the experimental results. Overall firing rates, properties of adaptation etc are unlikely to be very accurate.
One interesting preliminary observation has been that the effect of a neuron onto itself
(its receptor current lowering the LFP) could constitute in itself a form of adaptation.
However, it is unclear whether this can be made to work as it necessitates that the LFP
works on the (relatively slow) timescale of the observed adaptation. Bur something to keep an eye on.